test: ユニットテストファイルを分割して整理
「開く」ボタンのテストを、別ファイルに分割した。
